requirements.txt
# Core dependencies
google-cloud-vision>=3.0.0
Pillow>=9.0.0
numpy>=1.20.0
python-dateutil>=2.8.0

# Optional: Install with `pip install -e ".[ml]"` for ML features
# spacy>=3.0.0
# transformers>=4.20.0
# torch>=1.12.0
# scikit-learn>=1.0.0
